PayOS config

+++ Domain/Entities/PaymentTransaction.cs
@@ -9,7 +9,7 @@ namespace Domain.Entities
     public class PaymentTransaction
     {
         public Guid Id { get; set; }
-        public decimal Amount { get; set; }
+        public float Amount { get; set; }
         public string OrderId { get; set; }
         public string Status { get; set; }
         public DateTime CreatedAt { get; set; }
